+++ utils/src/time.rs
@@ -1,6 +1,6 @@
-pub fn now_sec() -> u32 {
+pub fn now_sec() -> u64 {
     std::time::SystemTime::now()
         .duration_since(std::time::UNIX_EPOCH)
         .unwrap()
-        .as_secs() as u32
+        .as_secs()
 }
